ended request to write back to caller
Without `req.response().end();` the event was hanging, i.e. if there were events every minute it was hanging on each for 5 minutes.

It is something very simple but in my case, I copied trusting that the code is correct and then had to debug for hours to find out the problem.

Signed-off-by: Rinor Maloku <rinormaloku37@gmail.com>